Where to begin... It might be best to back up anything you have been working on but didn't upload and then delete everything in the folder. Then reCheckout the project.
The process has been broken into multiple parts (since SVN is being a pain) but here is the full summary.

There is now a content project for SGDe, this allows for easier game development by separating the game from the content. Some other modifications include the removal of a previous change to the physics system where vectors were used. It may be cleaner code to multiply two Vector2s but it is faster to multiply the components and more memory efficient then creating a new Vector2 every time. Removal of CollisionBugDemo since the bug was fixed. Documentation for the engine and format (not complete but next on "the list") for the content. For the Demo project I removed the StaticBall since it the content project took care of nearly all operations that StaticBall was created for. Also a link to a C# tutorial and a (large) video tutorial on how to make a Tile Based Engine (not trying to push this but a great resource for learning XNA and C#).

Take a look at the Demo project's content folder to see examples of the content system. If you have any ideas or modifications or bugs, definitely bugs, let me know either in email/txt or at the next meeting. Sorry for the inconvenience of such a big change and thanks for the patience.

Update #1