An additional update (in addition to Brent's). Note: Both my update and Brent's can be summed up as "Stuff we worked on but never actually finished/fully-tested but we might as well upload them". In Brent's case, he didn't know but it seems to be minor (initial) optimization of the physics system (specifically scaling). Which he subsequently broke in SceneNode.
My update:
Update to the content format specification.
Support for map specific camera scale, rotation, and bounds.
Additional safety comparisions in PhysicsPharaoh.
Map specific physics (unsure how finished this is [update: seems to be ignored for now])
More support and implementation of the Date object.
Better reporting of errors in script system (which is NOT done so don't expect it to do anything)
General updates to comment remover for scripting system.